+++ packages/util-crypto/src/base58/index.ts
@@ -1,0 +1,10 @@
+// Copyright 2017-2020 @polkadot/util-crypto authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+/**
+ * @summary Encode and decode base58 values
+ */
+
+export { default as base58Decode } from './decode';
+export { default as base58Encode } from './encode';