+++ packages/util-crypto/src/index.ts
@@ -10,6 +10,7 @@ export * from './init';
 
 // all internal exports
 export * from './address';
+export * from './base58';
 export * from './blake2';
 export * from './keccak';
 export * from './key';